# swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
...
  "info": {
    "title": "skinswipe",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/support/rooms": {
      "get": {
        "description": "Список комнат",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/room/{roomName}": {
      "get": {
        "description": "getRoom",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "roomName",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/messages/{roomName}/{limit}/{offset}": {
      "get": {
        "description": "messages",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "roomName",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "limit",
            "type": "number",
            "required": true
          },
          {
            "in": "path",
            "name": "offset",
            "type": "number",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/resolve/{steamId}": {
      "get": {
        "description": "resolve",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "steamId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/banList": {
      "get": {
        "description": "banList",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/bookmarkList": {
      "get": {
        "description": "bookmarkList",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/ban/{steamId}/{hours}": {
      "get": {
        "description": "ban",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "steamId",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "hours",
            "type": "number",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/abuse/{steamId}": {
      "get": {
        "description": "abuse",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "steamId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/bookmark/{steamId}": {
      "get": {
        "description": "bookmark",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "steamId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/globalBan/{steamId}": {
      "get": {
        "description": "globalBan",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "steamId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/makeTop/{steamId}": {
      "get": {
        "description": "makeTop",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "steamId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/addCoins/{steamId}": {
      "get": {
        "description": "addCoins",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "steamId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/support/savePoints/{steamId}/{points}": {
      "get": {
        "description": "savePoints",
        "tags": ["support"],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "appVersion",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "steamId",
            "type": "string",
            "required": true
          },
          {
            "in": "path",
            "name": "points",
            "type": "number",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/translation/{screen}/{locale}/{platform}": {
      "get": {
        "description": "Получение текста для конкретного экрана",
        "tags": ["translation"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "screen",
            "type": "string",
            "required": true,
            "default": ""
          },
          {
            "in": "path",
            "name": "locale",
            "type": "string",
            "required": true,
            "default": ""
          },
          {
            "in": "path",
            "name": "platform",
            "type": "string",
            "required": true,
            "default": ""
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  }
}